!XI ina/master -> ina/malden
!D (Dialog) Remove the now unused dialog handler functions

!T (Dialogs) Replace dialog handlers to new gear dialog functions

!T (functions) Add the new gear dialog functions to config

!B (BaseTheme) Fix script errors when no base theme is defined
Also added a hint to the save dialog detailing that no theme is selected

!B (Configs) Removed config init check
Initial idea was to keep all config data on server only to save memory for clients.
This would require a client-server architecture refactor beyond the necessary scope of this project.
#Flag Remove this line entirely, or keep init duplication check.

!A (Dialogs) Added refactored gear dialog functions

!T (config) Renamed rocket ammo type from "launcher" to "rocket"
Allows to differentiate between weapons and munitions

!I (Configs) Integrate auto parsed define data into new config system
New config system now perfectly matches current data in gear defines for:
-Magazines
-Weapons
-Misc Items